+++ StabilityMatrix.Avalonia/ViewModels/Dialogs/LaunchOptionsViewModel.cs
@@ -1,9 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
+using System.ComponentModel;
 using System.Linq;
+using System.Reactive.Linq;
+using System.Threading;
 using CommunityToolkit.Mvvm.ComponentModel;
 using FuzzySharp;
+using Microsoft.Extensions.Logging;
 using StabilityMatrix.Avalonia.Views.Dialogs;
 using StabilityMatrix.Core.Attributes;
 using StabilityMatrix.Core.Helper.Cache;
@@ -14,26 +18,28 @@ namespace StabilityMatrix.Avalonia.ViewModels.Dialogs;
 [View(typeof(LaunchOptionsDialog))]
 public partial class LaunchOptionsViewModel : ContentDialogViewModelBase
 {
+    private readonly ILogger<LaunchOptionsViewModel> logger;
     private readonly LRUCache<string, ImmutableList<LaunchOptionCard>> cache = new(100);
-    
-    [ObservableProperty] private string title = "Launch Options";
-    [ObservableProperty] private bool isSearchBoxEnabled = true;
-    
+
     [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(FilteredCards))]
+    private string title = "Launch Options";
+
+    [ObservableProperty]
+    private bool isSearchBoxEnabled = true;
+
+    [ObservableProperty]
     private string searchText = string.Empty;
-    
-    [ObservableProperty] 
+
+    [ObservableProperty]
     private IReadOnlyList<LaunchOptionCard>? filteredCards;
-    
+
     public IReadOnlyList<LaunchOptionCard>? Cards { get; set; }
-    
+
     /// <summary>
     /// Return cards that match the search text
     /// </summary>
-    private IReadOnlyList<LaunchOptionCard>? GetFilteredCards()
+    private IReadOnlyList<LaunchOptionCard>? GetFilteredCards(string? text)
     {
-        var text = SearchText;
         if (string.IsNullOrWhiteSpace(text) || text.Length < 2)
         {
             return Cards;
@@ -50,18 +56,30 @@ public partial class LaunchOptionsViewModel : ContentDialogViewModelBase
             Type = LaunchOptionType.Bool,
             Options = Array.Empty<LaunchOption>()
         };
-            
-        var extracted = Process
-            .ExtractTop(searchCard, Cards, c => c.Title.ToLowerInvariant());
-        var results = extracted
-            .Where(r => r.Score > 40)
-            .Select(r => r.Value)
-            .ToImmutableList();
+
+        var extracted = Process.ExtractTop(searchCard, Cards, c => c.Title.ToLowerInvariant());
+        var results = extracted.Where(r => r.Score > 40).Select(r => r.Value).ToImmutableList();
         cache.Add(text, results);
         return results;
     }
 
-    public void UpdateFilterCards() => FilteredCards = GetFilteredCards();
+    public void UpdateFilterCards() => FilteredCards = GetFilteredCards(SearchText);
+
+    public LaunchOptionsViewModel(ILogger<LaunchOptionsViewModel> logger)
+    {
+        this.logger = logger;
+
+        Observable
+            .FromEventPattern<PropertyChangedEventArgs>(this, nameof(PropertyChanged))
+            .Where(x => x.EventArgs.PropertyName == nameof(SearchText))
+            .Throttle(TimeSpan.FromMilliseconds(50))
+            .Select(_ => SearchText)
+            .ObserveOn(SynchronizationContext.Current!)
+            .Subscribe(
+                text => FilteredCards = GetFilteredCards(text),
+                err => logger.LogError(err, "Error while filtering launch options")
+            );
+    }
 
     public override void OnLoaded()
     {
@@ -75,8 +93,9 @@ public partial class LaunchOptionsViewModel : ContentDialogViewModelBase
     public List<LaunchOption> AsLaunchArgs()
     {
         var launchArgs = new List<LaunchOption>();
-        if (Cards is null) return launchArgs;
-        
+        if (Cards is null)
+            return launchArgs;
+
         foreach (var card in Cards)
         {
             launchArgs.AddRange(card.Options);